Fix init script required package installation
Without the -y flag on 'apt install', the command requires confirmation
from the user. Unfortunately, the rest of the script can complete before
the input, meaning none of the packages get installed. It seems that the
confirmation input does not block the script, which is free to continue
in the background.