Track custom/ folder to avoid errors and provide scaffolding for user
Previously, using `,ve` in vim to edit the vimrcs would not write the
custom vimrc if the parent custom/ folder did not exist.
There are 2 solutions to this:
1. Force vim to create parent directory on write
2. Provide directory to user in git

I opted for the latter since use of the custom folder is expected and
it reduces the error involved with the user having to create it
themselves.

Note the .gitignore inside custom, which ignores all files within the
folder, ensuring the empty folder is tracked.